refactor: modulization
